requirements.txt
streamlit
pillow
pytesseract
python-docx
opencv-python-headless
numpy
//